External links

Contains information linking the entities to their original sources (including the IBERIAN subset (Harrison et al., 2022a), EMBSeCBIO (Harrison et al., 2021), NEOTOMA (Williams et al., 2018) and RPD (Harrison et al., 2022a)). Represents the external_link table in the database.

external_link

Format

A data frame (tibble object) with 1977 rows and 7 variables:

ID_SITE

Unsigned integer: Unique identifier for the site, multiple entities can be located within the same site

ID_ENTITY

Unsigned integer: Unique identifier for the entity

external_ID_SITE

Unsigned integer:

external_ID_ENTITY

Unsigned integer: Unique identifier for the entity

external_site_name

Text:

external_entity_name

Text:

external_source

Text:

References

Harrison, S.P., Marinova, E. and Cruz-Silva, E.: EMBSeCBIO pollen database. University of Reading. Dataset. doi: https://doi.org/10.17864/1947.309, 2021.

Harrison, S.P., Shen, Y. and Sweeney, L.: Pollen data and charcoal data of the Iberian Peninsula (version 3). University of Reading. Dataset. doi: https://doi.org/10.17864/1947.000369, 2022a.

Harrison, S.P., Villegas-Diaz, R., Lincoln, P., Kesner, D., Cruz-Silva, E., Sweeney, L., Shen, Y. and Gallagher, D.: The Reading Palaeofire Database v1b: an expanded global resource to document changes in fire regimes from sedimentary charcoal records. University of Reading. Dataset. doi: https://doi.org/10.17864/1947.000345, 2022b.

Williams, J.W., Grimm, E.G., Blois, J., Charles, D.F., Davis, E., Goring, S.J., Graham, R., Smith, A.J., Anderson, M., Arroyo-Cabrales, J., Ashworth, A.C., Betancourt, J.L., Bills, B.W., Booth, R.K., Buckland, P., Curry, B., Giesecke, T., Hausmann, S., Jackson, S.T., Latorre, C., Nichols, J., Purdum, T., Roth, R.E., Stryker, M., Takahara, H.: The Neotoma Paleoecology Database: A multi-proxy, international community-curated data resource. Quaternary Research 89, 156-177. http://www.neotomadb.org/, 2018.

Examples

external_link_tb <- special.epd::external_link
head(external_link_tb)
#> # A tibble: 6 × 7
#>   ID_SITE ID_ENTITY external_ID_SITE external_ID_ENTITY external_site_name
#>     <int>     <int>            <int>              <int> <chr>             
#> 1       1         1            27489              49347 Aalkistensee      
#> 2       2         2             1791               2116 Aammiq            
#> 3       2         2            27554              49501 Aammiq            
#> 4       3         3            27556              49505 Aansser peat bog  
#> 5       4         4            15736              24161 Abant Gölü        
#> 6       5         5             1342               1732 Abborrtjarnen     
#> # … with 2 more variables: external_entity_name <chr>, external_source <chr>